* Improved startup.  Moved special startup logic from StartupActivity to InheritingActivity.  Added check to avoid other apps setting scriptname.
